Improve KML parsing and region handling
- KML parser supports unnamed placemarks
- KML Parser
  - Add support for unnamed placemarks
  - Add MultiGeometry support
  - Improve region surface area display
- Region Handling
  - Warn if region spans >= 20,000 tiles
  - Display progress for tile enumeration
  - Only enumerate tiles once per command (removed `GetPolygonalRegionStats` which would also enumerate tiles)
- Fixed ear clipping triangulation algorithm
- Improve performance of `GeoPolygon.ContainsTile()`

+++ src/LibMapCommon/Geometry/TileStats.cs
@@ -8,4 +8,4 @@ public readonly record struct TileStats(
 	int MaxRow,
 	int MinColumn,
 	int MaxColumn,
-	int TileCount);
+	long TileCount);
